Time: 37 ms (25.68%), Space: 15.6 MB (45.68%) - LeetHub

+++ 2938-separate-black-and-white-balls/2938-separate-black-and-white-balls.cpp
@@ -5,7 +5,7 @@ public:
         int black = 0;
         for(int i=0; i<s.size(); i++){
             if(s[i] == '0')
-                swap += (long long) black;
+                swap +=  black;
             else
                 black++;
         }
